requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
python-dotenv>=0.19.0
streamlit
openai
